+++ database/migrations/2026_04_21_000006_create_bills_status_logs_table.php
@@ -10,10 +10,10 @@ return new class extends Migration
     {
         Schema::create('bills_status_logs', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('bill_id');
+            $table->unsignedInteger('bill_id');
             $table->tinyInteger('from_status')->nullable();
             $table->tinyInteger('to_status');
-            $table->unsignedBigInteger('operator_id')->nullable();
+            $table->unsignedInteger('operator_id')->nullable();
             $table->string('reason', 255)->nullable();
             $table->dateTime('created_at')->nullable();
